backend/requirements.txt
fastapi
uvicorn
pydantic
redis 
groq
python-multipart 
yt-dlp 
pymupdf
python-dotenv
youtube_transcript_api==1.0.3
google-api-python-client==2.100.0
requests
deepgram-sdk
supadata==1.1.0
sentence_transformers
PyPDF2
supabase
python-jose[cryptography]
numpy>=1.24.0
scikit-learn>=1.3.0
